[Manual] Remove message_type argument from get_message_collection
